Backend/requirements.txt
fastapi==0.104.0
uvicorn[standard]==0.23.2

pydantic==1.10.12
python-dotenv==1.0.0

pymongo==5.5.0
passlib[bcrypt]==1.7.5
bcrypt==4.1.2

apscheduler==3.11.1

maileroo==1.0.0
email-validator==2.1.1

python-multipart==0.0.6

requests==2.31.0

google-api-python-client==2.118.0
google-auth==2.26.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
